refactor to FSD, fix some errors
