+++ package.json
@@ -24,13 +24,14 @@
     "@eslint/js": "^9.29.0",
     "@tanstack/eslint-plugin-query": "^5.81.2",
     "@tanstack/router-plugin": "^1.121.29",
+    "@types/node": "^24.0.7",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
     "@typescript-eslint/eslint-plugin": "^8.34.1",
     "@vitejs/plugin-react-swc": "^3.9.0",
     "autoprefixer": "^10.4.21",
     "eslint": "^9.29.0",
-    "eslint-import-resolver-typescript": "^4.4.3",
+    "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-prettier": "^5.5.0",
     "eslint-plugin-react-dom": "^1.52.2",
