+++ src/app/layouts/RootLayout.tsx
@@ -2,9 +2,10 @@ import { HeroUIProvider, Link } from '@heroui/react';
 import { Outlet } from '@tanstack/react-router';
 import { TanStackRouterDevtools } from '@tanstack/react-router-devtools';
 
-import { useHref } from './hooks/useHeroHref';
-import { useHeroNavigate } from './hooks/useHeroNavigate';
-import { router } from './router';
+import { useHref } from '@/hooks/useHeroHref';
+import { useHeroNavigate } from '@/hooks/useHeroNavigate';
+
+import { router } from '../router/router';
 
 const RootLayout = () => {
   const navigate = useHeroNavigate(router);
