+++ src/app/router/router.tsx
@@ -1,0 +1,13 @@
+import { createRouter } from '@tanstack/react-router';
+
+import { routeTree } from './routes';
+
+export const router = createRouter({ routeTree });
+
+declare module '@tanstack/react-router' {
+  interface Register {
+    router: typeof router;
+  }
+}
+
+export type RouterType = typeof router;
