+++ src/hooks/useHeroHref.ts
@@ -1,6 +1,6 @@
 import { useLinkProps } from '@tanstack/react-router';
 
-import type { router } from '../router';
+import type { router } from '../app/router/router';
 
 type Routes = Parameters<typeof router.navigate>[0]['to'];
