+++ src/pages/About/index.tsx
@@ -1,0 +1,1 @@
+export const AboutPage = () => <div className="p-4">About Page</div>;
