+++ src/pages/Home/index.tsx
@@ -1,0 +1,1 @@
+export const HomePage = () => <div className="p-4">Home Page</div>;
