+++ vite.config.ts
@@ -1,7 +1,18 @@
+import { dirname, resolve } from 'path';
+import { fileURLToPath } from 'url';
+
 import react from '@vitejs/plugin-react-swc';
 import { defineConfig } from 'vite';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      '@': resolve(__dirname, 'src'),
+    },
+  },
 });
